refactor(csharp-di): extract shared auth property utility, remove unused parameters
- Extract duplicated auth property extraction logic (~200 lines) from both
  ServiceCollectionExtensionsGenerator and DependencyInjectionOptionsGenerator
  into shared authPropertyUtils.ts
- Remove unused _authProperties parameters from addParameterlessOverload,
  addActionOverload, and addConfigurationSectionOverload methods
- No behavioral changes; all tests pass
