+++ generators/csharp/sdk/src/di/DependencyInjectionOptionsGenerator.ts
@@ -1,13 +1,9 @@
 import { CSharpFile, FileGenerator } from "@fern-api/csharp-base";
 import { ast } from "@fern-api/csharp-codegen";
 import { RelativeFilePath } from "@fern-api/fs-utils";
-import { FernIr } from "@fern-fern/ir-sdk";
 
 import { SdkGeneratorContext } from "../SdkGeneratorContext.js";
-import { collectInferredAuthCredentials } from "../utils/inferredAuthUtils.js";
-
-type AuthScheme = FernIr.AuthScheme;
-type OAuthScheme = FernIr.OAuthScheme;
+import { extractAuthProperties } from "./authPropertyUtils.js";
 
 /**
  * Generates {ClientName}Options.cs - a configuration options class for DI registration.
@@ -65,42 +61,16 @@ export class DependencyInjectionOptionsGenerator extends FileGenerator<CSharpFil
     }
 
     private addAuthProperties(class_: ast.Class): void {
-        const seenNames = new Set<string>();
-
-        for (const scheme of this.context.ir.auth.schemes) {
-            for (const prop of this.getPropertiesFromAuthScheme(scheme)) {
-                if (!seenNames.has(prop.name)) {
-                    class_.addField({
-                        access: ast.Access.Public,
-                        name: prop.pascalName,
-                        type: this.Primitive.string.asOptional(),
-                        get: true,
-                        set: true,
-                        summary: prop.docs
-                    });
-                    seenNames.add(prop.name);
-                }
-            }
-        }
-
-        // Also include global headers as options properties
-        for (const header of this.context.ir.headers) {
-            if (header.valueType.type === "container" && header.valueType.container.type === "literal") {
-                continue; // Skip literal headers
-            }
-            const name = header.name.name.camelCase.safeName;
-            const pascalName = header.name.name.pascalCase.safeName;
-            if (!seenNames.has(name)) {
-                class_.addField({
-                    access: ast.Access.Public,
-                    name: pascalName,
-                    type: this.Primitive.string.asOptional(),
-                    get: true,
-                    set: true,
-                    summary: header.docs ?? `The ${name} header value.`
-                });
-                seenNames.add(name);
-            }
+        const properties = extractAuthProperties(this.context);
+        for (const prop of properties) {
+            class_.addField({
+                access: ast.Access.Public,
+                name: prop.pascalName,
+                type: this.Primitive.string.asOptional(),
+                get: true,
+                set: true,
+                summary: prop.docs
+            });
         }
     }
 
@@ -139,146 +109,5 @@ export class DependencyInjectionOptionsGenerator extends FileGenerator<CSharpFil
             summary: "The timeout for HTTP requests. Defaults to 30 seconds.",
             initializer: this.csharp.codeblock("TimeSpan.FromSeconds(30)")
         });
-    }
-
-    /**
-     * Converts an auth scheme into options properties.
-     * Mirrors the parameter extraction logic in RootClientGenerator.getParameterFromAuthScheme.
-     */
-    private getPropertiesFromAuthScheme(scheme: AuthScheme): Array<{ name: string; pascalName: string; docs: string }> {
-        if (scheme.type === "header") {
-            const name = scheme.name.name.camelCase.safeName;
-            const pascalName = scheme.name.name.pascalCase.safeName;
-            return [
-                {
-                    name,
-                    pascalName,
-                    docs: scheme.docs ?? `The ${name} to use for authentication.`
-                }
-            ];
-        } else if (scheme.type === "bearer") {
-            const name = scheme.token.camelCase.safeName;
-            const pascalName = scheme.token.pascalCase.safeName;
-            return [
-                {
-                    name,
-                    pascalName,
-                    docs: scheme.docs ?? `The ${name} to use for authentication.`
-                }
-            ];
-        } else if (scheme.type === "basic") {
-            const usernameName = scheme.username.camelCase.safeName;
-            const usernamePascal = scheme.username.pascalCase.safeName;
-            const passwordName = scheme.password.camelCase.safeName;
-            const passwordPascal = scheme.password.pascalCase.safeName;
-            return [
-                {
-                    name: usernameName,
-                    pascalName: usernamePascal,
-                    docs: `The ${usernameName} to use for authentication.`
-                },
-                {
-                    name: passwordName,
-                    pascalName: passwordPascal,
-                    docs: `The ${passwordName} to use for authentication.`
-                }
-            ];
-        } else if (scheme.type === "oauth") {
-            const oauth = this.context.getOauth();
-            if (oauth == null) {
-                return [];
-            }
-            const properties: Array<{ name: string; pascalName: string; docs: string }> = [
-                {
-                    name: "clientId",
-                    pascalName: "ClientId",
-                    docs: "The client ID for OAuth authentication."
-                },
-                {
-                    name: "clientSecret",
-                    pascalName: "ClientSecret",
-                    docs: "The client secret for OAuth authentication."
-                }
-            ];
-            // Include additional OAuth parameters (custom properties and scopes)
-            properties.push(...this.getOAuthAdditionalProperties(scheme));
-            return properties;
-        } else if (scheme.type === "inferred") {
-            const inferred = this.context.getInferredAuth();
-            if (inferred == null) {
-                return [];
-            }
-            return this.getInferredAuthProperties(inferred);
-        }
-        return [];
-    }
-
-    /**
-     * Gets additional OAuth properties from custom token endpoint properties and scopes.
-     * Mirrors the logic in RootClientGenerator.getOAuthAdditionalConstructorParams.
-     */
-    private getOAuthAdditionalProperties(
-        scheme: OAuthScheme
-    ): Array<{ name: string; pascalName: string; docs: string }> {
-        const properties: Array<{ name: string; pascalName: string; docs: string }> = [];
-        for (const customProperty of scheme.configuration.tokenEndpoint.requestProperties.customProperties ?? []) {
-            if (this.isLiteralTypeReference(customProperty.property.valueType)) {
-                continue;
-            }
-            const typeRef = this.context.csharpTypeMapper.convert({
-                reference: customProperty.property.valueType
-            });
-            if (typeRef.isOptional) {
-                continue;
-            }
-            const name = customProperty.property.name.name.camelCase.safeName;
-            const pascalName = customProperty.property.name.name.pascalCase.safeName;
-            properties.push({
-                name,
-                pascalName,
-                docs: `The ${name} for OAuth authentication.`
-            });
-        }
-        const scopes = scheme.configuration.tokenEndpoint.requestProperties.scopes;
-        if (scopes && !this.isLiteralTypeReference(scopes.property.valueType)) {
-            const typeRef = this.context.csharpTypeMapper.convert({
-                reference: scopes.property.valueType
-            });
-            if (!typeRef.isOptional) {
-                const name = scopes.property.name.name.camelCase.safeName;
-                const pascalName = scopes.property.name.name.pascalCase.safeName;
-                properties.push({
-                    name,
-                    pascalName,
-                    docs: `The ${name} for OAuth authentication.`
-                });
-            }
-        }
-        return properties;
-    }
-
-    /**
-     * Gets properties from inferred auth credentials.
-     * Mirrors the logic in RootClientGenerator.getParameterFromAuthScheme for inferred type.
-     */
-    private getInferredAuthProperties(
-        inferred: FernIr.InferredAuthScheme
-    ): Array<{ name: string; pascalName: string; docs: string }> {
-        const tokenEndpointReference = inferred.tokenEndpoint.endpoint;
-        const tokenEndpointHttpService = this.context.getHttpService(tokenEndpointReference.serviceId);
-        if (tokenEndpointHttpService == null) {
-            return [];
-        }
-        const tokenEndpoint = this.context.resolveEndpoint(tokenEndpointHttpService, tokenEndpointReference.endpointId);
-        const credentials = collectInferredAuthCredentials(this.context, tokenEndpoint);
-        return credentials.map((credential) => ({
-            name: credential.camelName,
-            pascalName: credential.pascalName,
-            docs: credential.docs ?? `The ${credential.camelName} for authentication.`
-        }));
-    }
-
-    private isLiteralTypeReference(typeReference: FernIr.TypeReference): boolean {
-        return typeReference.type === "container" && typeReference.container.type === "literal";
     }
 }
